--- modules/apps/portal-crypto-hash/portal-crypto-hash-provider/portal-crypto-hash-provider-message-digest/src/main/java/com/liferay/portal/crypto/hash/provider/message/digest/internal/MessageDigestCryptoHashProvider.java
@@ -1,68 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.crypto.hash.provider.message.digest.internal;
-
-import com.liferay.portal.crypto.hash.spi.CryptoHashProvider;
-import com.liferay.portal.crypto.hash.spi.CryptoHashProviderResponse;
-import com.liferay.portal.kernel.util.ArrayUtil;
-import com.liferay.portal.kernel.util.MapUtil;
-
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-
-import java.util.Collections;
-import java.util.Map;
-
-/**
- * @author Arthur Chan
- */
-public class MessageDigestCryptoHashProvider implements CryptoHashProvider {
-
-	public static final String ALGORITHM = "algorithm";
-
-	public MessageDigestCryptoHashProvider(String cryptoHashProviderFactoryName)
-		throws NoSuchAlgorithmException {
-
-		_cryptoHashProviderFactoryName = cryptoHashProviderFactoryName;
-
-		_messageDigest = MessageDigest.getInstance(_DEFAULT_ALGORITHM);
-	}
-
-	public MessageDigestCryptoHashProvider(
-			String cryptoHashProviderFactoryName,
-			Map<String, ?> cryptoHashProviderProperties)
-		throws NoSuchAlgorithmException {
-
-		_cryptoHashProviderFactoryName = cryptoHashProviderFactoryName;
-
-		_messageDigest = MessageDigest.getInstance(
-			MapUtil.getString(cryptoHashProviderProperties, ALGORITHM));
-	}
-
-	@Override
-	public CryptoHashProviderResponse generate(byte[] salt, byte[] input) {
-		return new CryptoHashProviderResponse(
-			_cryptoHashProviderFactoryName,
-			Collections.singletonMap(
-				"algorithm", _messageDigest.getAlgorithm()),
-			_messageDigest.digest(ArrayUtil.append(salt, input)));
-	}
-
-	private static final String _DEFAULT_ALGORITHM = "SHA-256";
-
-	private final String _cryptoHashProviderFactoryName;
-	private final MessageDigest _messageDigest;
-
-}
